requirements.txt
numpy==2.2.3
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
streamlit==1.37.1
torch==2.6.0
ultralytics==8.3.85
ultralytics==8.3.84
